[ios][camera] Allow opt out of barcode scanning deps (#40444)
# Why
Barcode scanning can add size to the final ipa which is not necessary if the user is not using the functionality. I'm providing a way to opt out of including the scanning dependencies. 

# How
Added `barcodeScannerEnabled` to the config plugin. We provide a stub module for barcode scanning when it is disabled which no-ops all the scanning functionality.

# Test Plan
Sandbox. Enabling and disabling works as expected.